add lib/processing/graphics.rb
