+++ lib/processing.rb
@@ -16,4 +16,5 @@ require 'processing/touch'
 require 'processing/shader'
 require 'processing/capture'
 require 'processing/graphics_context'
+require 'processing/graphics'
 require 'processing/processing'
